topic detection and polarity calculation/polarity/requirements.txt
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cymem==2.0.6
Flask==2.0.3
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
langcodes==3.3.0
MarkupSafe==2.1.0
murmurhash==1.0.6
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pathy==0.6.1
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
smart-open==5.2.1
spacy==3.2.2
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.13
tqdm==4.62.3
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.8
wasabi==0.9.0
Werkzeug==2.0.3
zope.event==4.5.0
zope.interface==5.4.0
